Clean up: get_public_key
